feat: Include esbuild so that Node functions work out of the box
The CDK `NodejsFunction` construct automatically bundles code via esbuild. However it requires esbuild to be installed. If it isn't installed, it will use Docker.

To make this work out of the box without requiring Docker, we can require esbuild.

CDK doesn't do that because their constructs are translated to other languages via JSII, so they can't rely on a full Node solution (even though it's a NodeJS-specific construct). At least that's what I read on GitHub.